requirements-test.txt
cssselect
factory_boy
flake8
ipdb
mock
pytest
pytest-bdd
pytest-django
pytest-mock
pytest-pythonpath
splinter